Rename and fix the "release" workflow to "Prepare Auto Release Branch" (#408)
* Clearify

* fix?

* fix

---------

+++ eng/publish.mjs
@@ -5,11 +5,19 @@ const branchName = "publish/auto-release";
 execSync(`git config --global user.email "autochangesetbot@microsoft.com"`);
 execSync(`git config --global user.name "Microsoft Auto Changeset Bot"`);
 execSync(`pnpm changeset version`);
-execSync(`git commit -am "Bump versions"`);
-execSync(`git push origin HEAD:${branchName} --force`);
+const stdout = execSync(`git status --porcelain`).toString();
 
-console.log();
-console.log("-".repeat(160));
-console.log("|  Link to create the PR");
-console.log(`|  https://github.com/Azure/cadl-ranch/pull/new/${branchName}  `);
-console.log("-".repeat(160));
+if (stdout.trim() !== "") {
+  console.log("Commiting the following changes:\n", stdout);
+
+  execSync(`git commit -am "Bump versions"`);
+  execSync(`git push origin HEAD:${branchName} --force`);
+
+  console.log();
+  console.log("-".repeat(160));
+  console.log("|  Link to create the PR");
+  console.log(`|  https://github.com/Azure/cadl-ranch/pull/new/${branchName}  `);
+  console.log("-".repeat(160));
+} else {
+  console.log("No changes to publish");
+}
